Store user UUID in Keychain instead of UserDefaults and make sure it's not migrated to different devices
